Remove api/Me in favor of server side actions
